refactor: rename drawer views for consistency
