+++ packages/desktop/components/buttons/DappsButton.svelte
@@ -2,13 +2,13 @@
     import { Icon } from '@ui'
     import { Icon as IconEnum } from '@auxiliary/icon'
     import { DrawerId, closeDrawer, drawerState, openDrawer } from '@desktop/auxiliary/drawer'
-    import { DappsDrawerRouter, dappsDrawerRouter } from '@views/dashboard/drawers/dapps'
+    import { DappConfigRouter, dappsDrawerRouter } from '@views/dashboard/drawers/dapp-config'
 
     function onDappsClick(): void {
         if ($drawerState.active) {
             closeDrawer()
         } else {
-            $dappsDrawerRouter = new DappsDrawerRouter()
+            $dappsDrawerRouter = new DappConfigRouter()
             openDrawer({ id: DrawerId.Dapps })
         }
     }
