+++ packages/desktop/views/dashboard/drawers/contact-book/views/ContactListDrawer.svelte
