+++ packages/desktop/views/dashboard/drawers/contact-book/views/RemoveContactDrawer.svelte
