+++ packages/desktop/views/dashboard/drawers/contact-book/views/RemoveNetworkAddressesDrawer.svelte
