+++ packages/desktop/views/dashboard/drawers/dapp-config/dapp-config-route.enum.ts
@@ -1,3 +1,3 @@
-export enum DappsDrawerRoute {
+export enum DappConfigRoute {
     ConnectedDapps = 'connectedDapps',
 }
