+++ packages/desktop/views/dashboard/drawers/dapp-config/index.ts
@@ -1,0 +1,4 @@
+export * from './dapp-config-route.enum'
+export * from './dapp-config.router'
+
+export { default as DappConfigRouterView } from './DappConfigRouterView.svelte'
