--- packages/desktop/views/dashboard/drawers/dapps/DappsRouterView.svelte
@@ -1,15 +1,0 @@
-<script lang="ts">
-    import { onDestroy } from 'svelte'
-    import { ConnectedDappsDrawer } from './drawers'
-    import { DappsDrawerRoute } from './dapps-route.enum'
-    import { dappsDrawerRoute, dappsDrawerRouter } from './dapps.router'
-
-    onDestroy(() => {
-        $dappsDrawerRouter.reset()
-        $dappsDrawerRouter = null
-    })
-</script>
-
-{#if $dappsDrawerRoute === DappsDrawerRoute.ConnectedDapps}
-    <ConnectedDappsDrawer drawerRouter={$dappsDrawerRouter} />
-{/if}
