--- packages/desktop/views/dashboard/drawers/dapps/index.ts
@@ -1,4 +1,0 @@
-export * from './dapps-route.enum'
-export * from './dapps.router'
-
-export { default as DappsRouterView } from './DappsRouterView.svelte'
