+++ packages/desktop/views/dashboard/drawers/network-config/index.ts
@@ -1,0 +1,4 @@
+export { default as NetworkConfigRouterView } from './NetworkConfigRouterView.svelte'
+
+export * from './network-config-route.enum'
+export * from './network-config.router'
