+++ packages/desktop/views/dashboard/drawers/network-config/network-config-route.enum.ts
