+++ packages/desktop/views/dashboard/drawers/network-config/network-config.router.ts
