+++ packages/desktop/views/dashboard/drawers/network-config/views/ChainInformationDrawer.svelte
