+++ packages/desktop/views/dashboard/drawers/network-config/views/ConnectLedgerDeviceDrawer.svelte
