+++ packages/desktop/views/dashboard/drawers/network-config/views/ConnectedChainsDrawer.svelte
