+++ packages/desktop/views/dashboard/drawers/network-config/views/RemoveChainDrawer.svelte
