+++ packages/desktop/views/dashboard/drawers/network-config/views/components/IscpChainInformation.svelte
