+++ packages/desktop/views/dashboard/drawers/network-config/views/components/index.ts
