+++ packages/desktop/views/dashboard/drawers/network-config/views/index.ts
